Rename EncryptForUseDefaultInsertColumnsTokenGenerator to EncryptInsertDefaultColumnsTokenGenerator and move package to insert
